fix file name & move functions to dedicated services
